Move method out of semantic tokens service
