+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Extensions/RazorCodeDocumentExtensions.cs
@@ -1,8 +1,12 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
+using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
 using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.AspNetCore.Razor.LanguageServer.Protocol;
+using Microsoft.CodeAnalysis.Razor;
+using Microsoft.CodeAnalysis.Razor.Workspaces.Extensions;
 
 namespace Microsoft.AspNetCore.Razor.LanguageServer.Extensions;
 
@@ -15,4 +19,50 @@ internal static class RazorCodeDocumentExtensions
             RazorLanguageKind.Html => document.GetHtmlDocument(),
             _ => throw new System.InvalidOperationException(),
         };
+
+    public static bool TryGetMinimalCSharpRange(this RazorCodeDocument codeDocument, Range razorRange, [NotNullWhen(true)] out Range? csharpRange)
+    {
+        SourceSpan? minGeneratedSpan = null;
+        SourceSpan? maxGeneratedSpan = null;
+
+        var sourceText = codeDocument.GetSourceText();
+        var textSpan = razorRange.ToTextSpan(sourceText);
+        var csharpDoc = codeDocument.GetCSharpDocument();
+
+        // We want to find the min and max C# source mapping that corresponds with our Razor range.
+        foreach (var mapping in csharpDoc.SourceMappings)
+        {
+            var mappedTextSpan = mapping.OriginalSpan.AsTextSpan();
+
+            if (textSpan.OverlapsWith(mappedTextSpan))
+            {
+                if (minGeneratedSpan is null || mapping.GeneratedSpan.AbsoluteIndex < minGeneratedSpan.Value.AbsoluteIndex)
+                {
+                    minGeneratedSpan = mapping.GeneratedSpan;
+                }
+
+                var mappingEndIndex = mapping.GeneratedSpan.AbsoluteIndex + mapping.GeneratedSpan.Length;
+                if (maxGeneratedSpan is null || mappingEndIndex > maxGeneratedSpan.Value.AbsoluteIndex + maxGeneratedSpan.Value.Length)
+                {
+                    maxGeneratedSpan = mapping.GeneratedSpan;
+                }
+            }
+        }
+
+        // Create a new projected range based on our calculated min/max source spans.
+        if (minGeneratedSpan is not null && maxGeneratedSpan is not null)
+        {
+            var csharpSourceText = codeDocument.GetCSharpSourceText();
+            var startRange = minGeneratedSpan.Value.ToRange(csharpSourceText);
+            var endRange = maxGeneratedSpan.Value.ToRange(csharpSourceText);
+
+            csharpRange = new Range { Start = startRange.Start, End = endRange.End };
+            Debug.Assert(csharpRange.Start.CompareTo(csharpRange.End) <= 0, "Range.Start should not be larger than Range.End");
+
+            return true;
+        }
+
+        csharpRange = null;
+        return false;
+    }
 }
